DPE-105 Adding pgbouncer.ini Config Management.  (#2)
[Jira Ticket](https://warthogs.atlassian.net/browse/DPE-105?atlOrigin=eyJpIjoiZmVkZjk3MTcxYzM0NGJkMDg1NTU0NTQ5YjY1YWVlNWIiLCJwIjoiaiJ9)

This PR implements a class in the pgb charm library that can handle pgbouncer.ini config files, including parsing, editing, and writing. The existing implementation is a placeholder that implements the bare minimum for the pgbouncer application to come online, whereas the proposed implementation is capable of making use of the entire [pgbouncer config spec](https://www.pgbouncer.org/config.html). 

This class implements a mutablemapping that's copied from a [ConfigParser](https://docs.python.org/3/library/configparser.html) object, allowing config options to be accessed and modified dictionary-style, as well as providing some basic validation. The reason this doesn't just inherit from a ConfigParser is that it only uses strings, and therefore can't represent some of the config values that are better represented as dicts or lists.

Changelog: 
- Added PgbIni class
  - Added the ability to parse strings that represent pgbouncer.ini files
  - Added the ability to pass in pgbouncer config as a dictionary 
- Replaced placeholder pgbouncer.ini generation

## Comments 

* adding pgbouncer.ini config management, including tests

* updated config representation

* added better PgbIni validation, and better docs

* finished validation and updated docstrings

* exhaustively testing pgbouncer.ini validation

* linting

* removed superfluous methods

* cleaning up docs and constructor

* cleaned up userlist parsing

* updated tests to check for reserved dbnames

* updates to tests, and copyright date

* responding to PR comments:
updated naming and docs to better describe functionality