DPE-136 Added Charm Config (#5)
## Issue
The charm has to expose some config options for pgbouncer to be configurable. This draft PR exists so we can discuss what is necessary and what is unnecessary, out of the proposed changes and the [pgbouncer config docs](https://www.pgbouncer.org/config.html)

Jira Ticket: [DPE-136](https://warthogs.atlassian.net/browse/DPE-136?atlOrigin=eyJpIjoiYmE3ZmJmMDJhODU0NGNlYjkyNmYxYTllYTk4NTJhMjMiLCJwIjoiaiJ9)

## Solution
This PR proposes a minimal config, whose options should provide enough information for the charm to calculate the rest of the config values that relate to pooling. The rest of the pgbouncer config should be defined as best practices by the charm, such as users and TLS settings. 

## Context
The defaults and documentation for these values are copied wholesale from the pgbouncer config docs. If this is unacceptable, please let me know so it can be changed. 

## Release Notes
- Adds a subset of pgbouncer config values to charm config. 

## Commits

* added preliminary charm config

* simplified logging and TLS config

* adding updated config

From these two values and the charm setup, we can generate a lot of the required config. The rest, we intend to define as part of the charm.

* adding admin-user back in

* updated README to include config options

* fixing markdown mistakes

* updated readme to better document generated values

* formatting

* updated docs