Fix Release Pipeline (#38)
# PR
## Proposal
Update pipeline runner to jammy so charmcraft can build the charm. 

## Context
Currently, the charm [can't be released](https://github.com/canonical/pgbouncer-operator/actions/runs/3487511396/jobs/5838224654) because there's no good base on which to build the charm. Building a charm on jammy seems to require a jammy substrate, so I've updated the CI to use 22.04

## Testing
- I've tested this locally, and I'm currently testing building and running a jammy charm in a 20.04 VM.  
- I'm struggling on how to verify the release action locally, so if anyone has any ideas I'd appreciate it :) 

## Commits
* updated pipeline runner to jammy

* updated build and run lxd versions

* reverted build to 22.04